move Model to its own file
